7.5 Удалит тестовые данные и правит build (ругается на AxiosRequestConfig, перезапишет как InternalAxiosRequestConfig).
