--- src/mocks/offers.ts
@@ -1,234 +1,0 @@
-import {Offer} from '../types/offer';
-
-export const offers: Offer[] = [
-  {
-    'id': '8de24f2a-8c8d-4045-8c09-39b5dd725345',
-    'title': 'Wood and stone place',
-    'type': 'house',
-    'price': 660,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/19.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.563341,
-      'longitude': 10.019654000000001,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 3.1
-  },
-  {
-    'id': '0081594e-81ce-48ea-a73e-566852a1b6f1',
-    'title': 'Canal View Prinsengracht',
-    'type': 'house',
-    'price': 308,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/9.jpg',
-    'city': {
-      'name': 'Brussels',
-      'location': {
-        'latitude': 50.846557,
-        'longitude': 4.351697,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.854557,
-      'longitude': 4.364697,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 3.1
-  },
-  {
-    'id': 'cea2eda2-93d3-430a-8757-47567f8cc341',
-    'title': 'Tile House',
-    'type': 'apartment',
-    'price': 452,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/8.jpg',
-    'city': {
-      'name': 'Cologne',
-      'location': {
-        'latitude': 50.938361,
-        'longitude': 6.959974,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.945361,
-      'longitude': 6.962974,
-      'zoom': 16
-    },
-    'isFavorite': true,
-    'isPremium': true,
-    'rating': 1.3
-  },
-  {
-    'id': 'd43822af-5204-498f-b3f6-8656f78b22b0',
-    'title': 'House in countryside',
-    'type': 'house',
-    'price': 834,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/16.jpg',
-    'city': {
-      'name': 'Cologne',
-      'location': {
-        'latitude': 50.938361,
-        'longitude': 6.959974,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.917361,
-      'longitude': 6.977974,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 2.4
-  },
-  {
-    'id': 'a75224b0-bb6a-4c09-8269-18379944c9e8',
-    'title': 'Perfectly located Castro',
-    'type': 'room',
-    'price': 140,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/13.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.868610000000004,
-      'longitude': 2.342499,
-      'zoom': 16
-    },
-    'isFavorite': true,
-    'isPremium': false,
-    'rating': 2.5
-  },
-  {
-    'id': '46a23c22-acb3-4191-a127-b0af188bf331',
-    'title': 'Canal View Prinsengracht',
-    'type': 'room',
-    'price': 209,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/2.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.858610000000006,
-      'longitude': 2.330499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 3.8
-  },
-  {
-    'id': '314769d4-5f2e-4cdb-998e-d7e38c173d34',
-    'title': 'Wood and stone place',
-    'type': 'room',
-    'price': 227,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/3.jpg',
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.37454,
-        'longitude': 4.897976,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 52.3909553943508,
-      'longitude': 4.85309666406198,
-      'zoom': 16
-    },
-    'isFavorite': true,
-    'isPremium': false,
-    'rating': 4.1
-  },
-  {
-    'id': '3a40cafb-015e-4323-92f4-c94f1fb7f679',
-    'title': 'Tile House',
-    'type': 'hotel',
-    'price': 202,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/9.jpg',
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.37454,
-        'longitude': 4.897976,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 52.3609553943508,
-      'longitude': 4.85309666406198,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 2
-  },
-  {
-    'id': '56bfb061-f365-4e4a-bcf7-8128243e6bf3',
-    'title': 'Amazing and Extremely Central Flat',
-    'type': 'hotel',
-    'price': 217,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/13.jpg',
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.37454,
-        'longitude': 4.897976,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 52.3909553943508,
-      'longitude': 4.929309666406198,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 4
-  },
-  {
-    'id': '2d24a23e-50f3-4c92-a2bf-b5d4db91e541',
-    'title': 'The house among olive ',
-    'type': 'hotel',
-    'price': 171,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/3.jpg',
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.37454,
-        'longitude': 4.897976,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 52.3809553943508,
-      'longitude': 4.939309666406198,
-      'zoom': 16
-    },
-    'isFavorite': true,
-    'isPremium': true,
-    'rating': 5
-  }
-];
