+++ src/services/api.ts
@@ -1,4 +1,4 @@
-import axios, {AxiosInstance, AxiosRequestConfig} from 'axios';
+import axios, {AxiosInstance, InternalAxiosRequestConfig} from 'axios';
 import {getToken} from './token';
 
 const BACKEND_URL = 'https://15.design.htmlacademy.pro/six-cities';
@@ -11,7 +11,7 @@ export const createAPI = (): AxiosInstance => {
   });
 
   api.interceptors.request.use(
-    (config: AxiosRequestConfig) => {
+    (config: InternalAxiosRequestConfig) => {
       const token = getToken();
 
       if (token && config.headers) {
