+++ src/store/action.ts
@@ -1,7 +1,9 @@
 import {createAction} from '@reduxjs/toolkit';
 import {City, Offer} from '../types/offer';
+import {Review} from '../types/review';
 
 const changeActiveCity = createAction<{activeCity: City}>('changeActiveCity');
 const initOffers = createAction<{offers: Offer[]}>('initOffers');
+const initReviews = createAction<{reviews: Review[]}>('initReviews');
 
-export {changeActiveCity, initOffers};
+export {changeActiveCity, initOffers, initReviews};
