+++ src/store/reducer.ts
@@ -1,16 +1,19 @@
 import {createReducer} from '@reduxjs/toolkit';
-import {changeActiveCity, initOffers} from './action';
+import {changeActiveCity, initOffers, initReviews} from './action';
 import {City, Offer} from '../types/offer';
-import {CITIES} from '../mocks/cities';
+import {Review} from '../types/review';
+import {CITIES} from '../const';
 
 type StateType = {
 activeCity: City;
 offers: Offer[];
+reviews: Review[];
 };
 
 const initialState: StateType = {
   activeCity: CITIES[0],
   offers: [],
+  reviews: [],
 };
 
 const reducer = createReducer(initialState, (builder) => {
@@ -22,6 +25,10 @@ const reducer = createReducer(initialState, (builder) => {
     .addCase(initOffers, (state, action) => {
       const {offers} = action.payload;
       state.offers = offers;
+    })
+    .addCase(initReviews, (state, action) => {
+      const {reviews} = action.payload;
+      state.reviews = reviews;
     });
 });
